fix(shift): shorten the reminder date on the Remind button
The button label carries the reminder date, which is wanted, but the
abbreviated month name made it long on a phone row and longer still in
German. A numeric month with no year keeps the same information in about
half the width: "Reminded on 9/14, 2:15 PM", "Erinnert am 14.9., 14:15".

The hover tooltip is built from the same values and is shortened with it
so the two cannot disagree.

+++ apps/frontend/src/domain/shift/components/shift-instance-volunteers-panel.tsx
@@ -213,7 +213,7 @@ export function ShiftInstanceVolunteersPanel({
                 ? t('inviteStatus.actionRemind')
                 : t('inviteStatus.remindedAtTooltip', {
                     when: `${formatDate(new Date(invite.remindedAt as string), {
-                      month: 'short',
+                      month: 'numeric',
                       day: 'numeric',
                     })}, ${formatTime(new Date(invite.remindedAt as string))}`,
                   }),
@@ -225,7 +225,7 @@ export function ShiftInstanceVolunteersPanel({
           ? {
               Remind: t('inviteStatus.remindedAtTooltip', {
                 when: `${formatDate(new Date(invite.remindedAt), {
-                  month: 'short',
+                  month: 'numeric',
                   day: 'numeric',
                 })}, ${formatTime(new Date(invite.remindedAt))}`,
               }),
